common-exception-spring-boot-autoconfigure/src/main/java/org/bremersee/exception/spring/boot/autoconfigure/reactive/DefaultWebClientErrorDecoderAutoConfiguration.java
/*
 * Copyright 2020-2022 the original author or authors.
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *      http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */

package org.bremersee.exception.spring.boot.autoconfigure.reactive;

import com.fasterxml.jackson.databind.ObjectMapper;
import lombok.extern.slf4j.Slf4j;
import org.bremersee.exception.RestApiExceptionParser;
import org.bremersee.exception.spring.boot.autoconfigure.RestApiExceptionParserAutoConfiguration;
import org.bremersee.exception.webclient.DefaultWebClientErrorDecoder;
import org.springframework.beans.factory.ObjectProvider;
import org.springframework.boot.autoconfigure.AutoConfigureAfter;
import org.springframework.boot.autoconfigure.condition.ConditionalOnClass;
import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
import org.springframework.boot.autoconfigure.condition.ConditionalOnWebApplication;
import org.springframework.boot.autoconfigure.condition.ConditionalOnWebApplication.Type;
import org.springframework.boot.context.event.ApplicationReadyEvent;
import org.springframework.context.annotation.Bean;
import org.springframework.context.annotation.Configuration;
import org.springframework.context.event.EventListener;
import org.springframework.util.ClassUtils;

/**
 * The default web client error decoder auto configuration.
 *
 * @author Christian Bremer
 */
@ConditionalOnWebApplication(type = Type.REACTIVE)
@ConditionalOnClass({
    ObjectMapper.class,
    DefaultWebClientErrorDecoder.class
})
@AutoConfigureAfter({
    RestApiExceptionParserAutoConfiguration.class
})
@Configuration
@Slf4j
public class DefaultWebClientErrorDecoderAutoConfiguration {

  /**
   * Init.
   */
  @EventListener(ApplicationReadyEvent.class)
  public void init() {
    log.info("\n"
            + "*********************************************************************************\n"
            + "* {}\n"
            + "*********************************************************************************",
        ClassUtils.getUserClass(getClass()).getSimpleName());
  }

  /**
   * Default web client error decoder default web client error decoder.
   *
   * @param restApiExceptionParser the rest api exception parser
   * @return the default web client error decoder
   */
  @ConditionalOnMissingBean
  @Bean
  public DefaultWebClientErrorDecoder defaultWebClientErrorDecoder(
      ObjectProvider<RestApiExceptionParser> restApiExceptionParser) {

    return new DefaultWebClientErrorDecoder(restApiExceptionParser.getIfAvailable());
  }

}
